Add files via upload
To make the IPL match prediction capabilities accessible to a wider audience, the trained machine learning model was integrated into a Streamlit-based web application. The web application provides a user-friendly interface where users can input the live match details, such as the teams playing, the current score, and the number of wickets lost.